requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.15
exceptiongroup==1.2.2
fastapi==0.115.5
googleapis-common-protos==1.59.1
greenlet==3.1.1
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
importlib_metadata==8.5.0
Mako==1.3.6
MarkupSafe==3.0.2
opentelemetry-api==1.28.2
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-instrumentation-requests==0.49b2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
packaging==24.2
prometheus_client==0.21.0
protobuf==4.25.5
pydantic==2.10.2
pydantic_core==2.27.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
thrift==0.21.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
wrapt==1.17.0
zipp==3.21.0
